logging initiatve: skip stdout/stderr loggers
Keep track of all ::Logger instances that are attempting to write. If
there is a non STDOUT/STDERR based logger that is known, then mark all
STDOUT/STDERR based loggers to be skipped and not have their output
reported.

This is a safe keeping copy with tweaks added for PR #1041 / Issue
 #1039. Feel Free to delete this branch once the issue is confirmed to
not need to be addressed.